+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 - Delete deprecated `p[metric|log|trace].MarshalerSizer` interfaces (#6083)
 
+### 💡 Enhancements 💡
+
+- Add AppendEmpty and EnsureCapacity method to primitive pdata slices (#6060)
 
 ## v0.60.0 Beta